Move GetLatestRelease() to pkg/version
