+++ pkg/standalone/standalone.go
@@ -10,7 +10,6 @@ import (
 	"archive/zip"
 	"compress/gzip"
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
@@ -31,6 +30,7 @@ import (
 
 	"github.com/briandowns/spinner"
 	"github.com/dapr/cli/pkg/print"
+	pkg_version "github.com/dapr/cli/pkg/version"
 	"github.com/dapr/cli/utils"
 )
 
@@ -112,44 +112,6 @@ func Init(runtimeVersion string, dockerNetwork string, installLocation string) e
 	}
 
 	return nil
-}
-
-// nolint:gosec
-// GetLatestRelease return the latest release version of dapr
-func GetLatestRelease(gitHubOrg, gitHubRepo string) (string, error) {
-	releaseURL := fmt.Sprintf("https://api.github.com/repos/%s/%s/releases", gitHubOrg, gitHubRepo)
-	resp, err := http.Get(releaseURL)
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != 200 {
-		return "", fmt.Errorf("%s - %s", releaseURL, resp.Status)
-	}
-
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return "", err
-	}
-
-	var githubRepoReleases []githubRepoReleaseItem
-	err = json.Unmarshal(body, &githubRepoReleases)
-	if err != nil {
-		return "", err
-	}
-
-	if len(githubRepoReleases) == 0 {
-		return "", fmt.Errorf("no releases")
-	}
-
-	for _, release := range githubRepoReleases {
-		if !strings.Contains(release.TagName, "-rc") {
-			return release.TagName, nil
-		}
-	}
-
-	return "", fmt.Errorf("no releases")
 }
 
 func isDockerInstalled() bool {
@@ -319,7 +281,7 @@ func installDaprBinary(wg *sync.WaitGroup, errorChan chan<- error, dir, version 
 
 	if version == daprLatestVersion {
 		var err error
-		version, err = GetLatestRelease(DaprGitHubOrg, DaprGitHubRepo)
+		version, err = pkg_version.GetLatestRelease()
 		if err != nil {
 			errorChan <- fmt.Errorf("cannot get the latest release version: %s", err)
 			return
@@ -530,13 +492,6 @@ func moveFileToPath(filepath string, installLocation string) (string, error) {
 	}
 
 	return destFilePath, nil
-}
-
-type githubRepoReleaseItem struct {
-	URL     string `json:"url"`
-	TagName string `json:"tag_name"`
-	Name    string `json:"name"`
-	Draft   bool   `json:"draft"`
 }
 
 // nolint:gosec
