+++ pkg/version/version.go
@@ -6,8 +6,13 @@
 package version
 
 import (
+	"encoding/json"
+	"fmt"
+	"io/ioutil"
+	"net/http"
 	"os/exec"
 	"runtime"
+	"strings"
 )
 
 // GetRuntimeVersion returns the version for the local Dapr runtime.
@@ -25,4 +30,49 @@ func GetRuntimeVersion() string {
 	}
 	return string(out)
 
+}
+
+type githubRepoReleaseItem struct {
+	URL     string `json:"url"`
+	TagName string `json:"tag_name"`
+	Name    string `json:"name"`
+	Draft   bool   `json:"draft"`
+}
+
+// nolint:gosec
+// GetLatestRelease return the latest release version of dapr
+func GetLatestRelease(gitHubOrg, gitHubRepo string) (string, error) {
+	releaseURL := fmt.Sprintf("https://api.github.com/repos/%s/%s/releases", gitHubOrg, gitHubRepo)
+	resp, err := http.Get(releaseURL)
+	if err != nil {
+		return "", err
+	}
+	defer resp.Body.Close()
+
+	if resp.StatusCode != 200 {
+		return "", fmt.Errorf("%s - %s", releaseURL, resp.Status)
+	}
+
+	body, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return "", err
+	}
+
+	var githubRepoReleases []githubRepoReleaseItem
+	err = json.Unmarshal(body, &githubRepoReleases)
+	if err != nil {
+		return "", err
+	}
+
+	if len(githubRepoReleases) == 0 {
+		return "", fmt.Errorf("no releases")
+	}
+
+	for _, release := range githubRepoReleases {
+		if !strings.Contains(release.TagName, "-rc") {
+			return release.TagName, nil
+		}
+	}
+
+	return "", fmt.Errorf("no releases")
 }
